Implemented setting the links in the event info display
When you click on an event, the "side panel" is updated with the
event info, and with three buttons.  Guest list signup, reservation
signup, and ticket sales.  Those buttons are revealed or hidden
based on the information for the event, from the API response.

I implemented the show/hide of the buttons, and the link targets
that they point to.  For example, when a tickets URL is set, that
URL will be used.